video/drm: display: Check &display_state.is_init before display_bmp()
The commit to be fixed is not well thought out. It just take into
account the scenario where the U-Boot logo is already displayed and
the specific BMP is showed via the cmd 'rockchip_show_bmp'.

For the charge animation driver, the rockchip_show_bmp() will be
called without rockchip_show_logo() invoked, which means the
&display_state.is_init is false. It makes the charge BMPs unable to
be displayed as expected.

Therefore, we add the &display_state.is_init check before
display_bmp() in API rockchip_show_bmp(). It helps confirm that the
API rockchip_show_bmp() can work well as before.

Fixes: fc5f6aa8274 ("video/drm: display: Use new display_bmp() instead of display_logo for command rockchip_show_bmp")
Change-Id: I66688c0620443c086222d73fbb9d8602d601af30
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>