drivers: video: ebc: Continue driving sdclk at the end of the data line for rk3576
Some EINK panels require dummy clocks to be inserted at the end of the
data line. This means ebc need to keep driving the sdclk during LEL.

Some EINK panels do not specify whether dummy clocks need to be
inserted, but the test results show that there is no effect on
the display regardless of whether they are inserted.

Change-Id: I17db5ccef182d4702bd40cbf8960315406120443
Signed-off-by: Zorro Liu <lyx@rock-chips.com>