video/drm: analogix_dp: Set the hot plug deglitch time to 2ms
As described in Section 5.2.4 of the DP v1.2a standard:

A Hot Plug event, detected by an upstream DP device as a long HPD
pulse (wider than 2ms), prompts the DP Source device to take many
actions such as reading of EDID and RX capability in DPCD, and the
communication to its operating system.

The hot plug deglitch time should be set to 2ms.

Change-Id: I3ec50011442690b8dfbaa85fe544ec03ecc9e15b
Signed-off-by: Damon Ding <damon.ding@rock-chips.com>